requirements.txt
Flask==2.2.2
pydantic==1.10.2
SQLAlchemy==1.4.44
psycopg2-binary==2.9.5
flask-bcrypt==1.0.1
python-dotenv==0.21.0